Fix build to work with current codegangsta/cli
Name structs so we ignore new (unused) values in the codegangsta StringFlag
struct (in this case they introduced a Hidden bool which broke our
implicitly-constructed structs)

Also cmd.Action was changed from a func(*cli.Context) to a interface{}, so we
cast it to a func(*cli.Context) before attempting to take extra action on it.

I *think* these changes are backwards compatible, in any case it builds now.